23_03_15 10:28

+++ src/main/java/teamPro/bbangShuttle/controller/ItemController.java
@@ -4,15 +4,13 @@ import io.jsonwebtoken.Jwts;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.util.StringUtils;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 import teamPro.bbangShuttle.dto.ItemDTO;
-import teamPro.bbangShuttle.security.TokenProvider;
 import teamPro.bbangShuttle.service.ItemService;
 import teamPro.bbangShuttle.service.ReviewService;
 import teamPro.bbangShuttle.vo.ItemVO;
 
-import javax.servlet.http.HttpServletRequest;
 
 @RestController
 @RequiredArgsConstructor
@@ -20,13 +18,27 @@ import javax.servlet.http.HttpServletRequest;
 public class ItemController {
     private final ItemService itemService;
     private final ReviewService reviewService;
-    private final TokenProvider tokenProvider;
 
     @GetMapping
     public ResponseEntity<?> itemList() {
         try {
             ItemDTO<ItemVO> response = ItemDTO.<ItemVO>builder()
                     .itemList(itemService.findAllItem())
+                    .build();
+            return ResponseEntity.ok().body(response);
+        } catch (Exception e) {
+            ItemDTO<ItemVO> response = ItemDTO.<ItemVO>builder()
+                    .error(e.getMessage())
+                    .build();
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(response);
+        }
+    }
+
+    @GetMapping("/home")
+    public ResponseEntity<?> itemHomeList() {
+        try {
+            ItemDTO<ItemVO> response = ItemDTO.<ItemVO>builder()
+                    .itemList(itemService.randList())
                     .build();
             return ResponseEntity.ok().body(response);
         } catch (Exception e) {
@@ -54,11 +66,11 @@ public class ItemController {
     }
 
     @PutMapping
-    public ResponseEntity<?> itemSave(@RequestBody ItemVO vo, HttpServletRequest request) {
+    public ResponseEntity<?> itemSave(@RequestBody ItemVO vo, @AuthenticationPrincipal String userID) {
         try {
-//            if (!"admin".equals(tokenProvider.validateAndGetUserId(getTokenFromRequest(request)))) {
-//                throw new IllegalArgumentException("Unauthorized access");
-//            }
+            if (!"admin".equals(userID)) {
+                throw new IllegalArgumentException("관리자 권한이 없습니다.");
+            }
             if (itemService.save(vo) < 1) {
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
             }
@@ -75,11 +87,11 @@ public class ItemController {
     }
 
     @DeleteMapping
-    public ResponseEntity<?> itemDelete(@RequestBody ItemVO vo,HttpServletRequest request) {
+    public ResponseEntity<?> itemDelete(@RequestBody ItemVO vo,@AuthenticationPrincipal String userID) {
         try {
-//            if (!"admin".equals(tokenProvider.validateAndGetUserId(getTokenFromRequest(request)))) {
-//                throw new IllegalArgumentException("Unauthorized access");
-//            }
+            if (!"admin".equals(userID)) {
+                throw new IllegalArgumentException("Unauthorized access");
+            }
             if (itemService.delete(vo) < 1) {
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
             }
@@ -96,11 +108,11 @@ public class ItemController {
     }
 
     @PatchMapping
-    public ResponseEntity<?> itemUpdate(@RequestBody ItemVO vo, HttpServletRequest request) {
+    public ResponseEntity<?> itemUpdate(@RequestBody ItemVO vo, @AuthenticationPrincipal String userID) {
         try {
-//            if (!"admin".equals(tokenProvider.validateAndGetUserId(getTokenFromRequest(request)))) {
-//                throw new IllegalArgumentException("Unauthorized access");
-//            }
+            if (!"admin".equals(userID)) {
+                throw new IllegalArgumentException("Unauthorized access");
+            }
             if (itemService.update(vo) < 1) {
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
             }
@@ -114,12 +126,5 @@ public class ItemController {
                     .build();
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(response);
         }
-    }
-    private String getTokenFromRequest(HttpServletRequest request) {
-        String bearerToken = request.getHeader("Authorization");
-        if (!StringUtils.isEmpty(bearerToken) && bearerToken.startsWith("Bearer ")) {
-            return bearerToken.substring(7);
-        }
-        return null;
     }
 }
